More work with the event bus. See description
Found a couple of functions where this is useful, like with the DisplayTutorial function in TutorialManager. However, since for stuff like PlaySFXAtPlayer we need to access the SoundManager script anyway to get the clip information, not much point in using the event bus as we already have to have a reference to the SoundManager script anyway so might as well use it for the function also.